bak/requirements.txt
aniso8601==9.0.1
appnope==0.1.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
boto==2.49.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.2
cssselect==1.1.0
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-labs==1.0.3
dash-table==5.0.0
dataclasses==0.8
decorator==5.1.1
execnet==1.9.0
fasteners==0.17.3
filelock==3.6.0
Flask==2.1.1
Flask-Compress==1.11
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.32.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
more-itertools==8.12.0
nose==1.3.7
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.2
parameterized==0.8.1
parso==0.8.3
path==16.4.0
pdfminer.six==20220319
percy==2.0.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==22.0.4
platformdirs==2.5.1
plotly==5.7.0
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyotp==2.6.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.1
pytest-forked==1.4.0
pytest-html==2.0.1
pytest-metadata==2.0.1
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rich==12.2.0
sbvirtualdisplay==1.1.0
scipy==1.8.0
selenium==4.1.3
setuptools==62.1.0
setuptools-scm==6.4.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
SQLAlchemy==1.4.35
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.9
waitress==2.1.1
wcwidth==0.2.5
webdriver-manager==3.5.4
Werkzeug==2.1.1
wheel==0.37.1
wsproto==1.1.0
zipp==3.8.0
